requirements.txt
Django==3.0.8
django-db-logger==0.1.9
djangorestframework==3.11.0
Pillow==7.2.0

